refactor chat types
