+++ GUI/src/types/chat.ts
@@ -55,3 +55,13 @@ export interface Chat {
   comment: string;
   labels: string;
 }
+
+export enum MessageSseEvent {
+  READ = 'message-read',
+  DELIVERED = 'message-delivered'
+}
+
+export type MessageStatus = {
+  messageId: string | null;
+  readTime: any;
+}
